# docs/openapi/openapi.yaml
...
openapi: 3.0.4
...
components:
  schemas:
...
    Instant:
      format: date-time
      type: string
      example: 2022-03-10T16:15:50Z
...
    TripReconstructionRequestDto:
      required:
      - segments
      type: object
      properties:
        tripId:
          format: int64
          type: integer
        segments:
          minItems: 1
          type: array
          items:
            $ref: "#/components/schemas/TripReconstructionSegmentDto"
    TripReconstructionSegmentDto:
      required:
      - segmentType
      - startTime
      - endTime
      type: object
      properties:
        segmentType:
          pattern: \S
          type: string
        startTime:
          $ref: "#/components/schemas/Instant"
        endTime:
          $ref: "#/components/schemas/Instant"
        locationName:
          maxLength: 255
          type: string
        latitude:
          format: double
          maximum: 90.0
          minimum: -90.0
          type: number
        longitude:
          format: double
          maximum: 180.0
          minimum: -180.0
          type: number
        movementType:
          type: string
        waypoints:
          type: array
          items:
            $ref: "#/components/schemas/TripReconstructionWaypointDto"
    TripReconstructionWaypointDto:
      required:
      - latitude
      - longitude
      type: object
      properties:
        latitude:
          format: double
          maximum: 90.0
          minimum: -90.0
          type: number
        longitude:
          format: double
          maximum: 180.0
          minimum: -180.0
          type: number
    TripStatus:
      enum:
      - UNPLANNED
      - UPCOMING
      - ACTIVE
      - COMPLETED
      - CANCELLED
      type: string
...
    UpdateTripDto:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 150
          pattern: \S
          type: string
        startTime:
          $ref: "#/components/schemas/Instant"
        endTime:
          $ref: "#/components/schemas/Instant"
        color:
          pattern: "^#[0-9A-Fa-f]{6}$"
          type: string
        notes:
          maxLength: 4000
          type: string
        status:
          $ref: "#/components/schemas/TripStatus"
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
...
  /api/reconstruction/commit:
    post:
      operationId: ReconstructionResource_commit
      tags:
      - "User: Trips and Planning"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripReconstructionRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
        "400":
          description: Bad Request
      summary: Commit
      security:
      - jwt:
        - USER
        - ADMIN
  /api/reconstruction/preview:
    post:
      operationId: ReconstructionResource_preview
      tags:
      - "User: Trips and Planning"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripReconstructionRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
        "400":
          description: Bad Request
      summary: Preview
      security:
      - jwt:
        - USER
        - ADMIN
...
  /api/trips/plan-suggestion:
    get:
      operationId: PlanSuggestionResource_getPlanSuggestion
      tags:
      - "User: Trips and Planning"
      parameters:
      - name: lat
        in: query
        schema:
          format: double
          type: number
      - name: lon
        in: query
        schema:
          format: double
          type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Get Plan Suggestion
      security:
      - jwt:
        - USER
        - ADMIN
  /api/trips/{id}:
    put:
      operationId: TripResource_updateTrip
      tags:
      - "User: Trips"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
        "400":
          description: Bad Request
      summary: Update Trip
      security:
      - jwt:
        - USER
        - ADMIN
    get:
      operationId: TripResource_getTrip
      tags:
      - "User: Trips"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Get Trip
      security:
      - jwt:
        - USER
        - ADMIN
    delete:
      operationId: TripResource_deleteTrip
      tags:
      - "User: Trips"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: mode
        in: query
        schema:
          default: unlink_only
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Delete Trip
      security:
      - jwt:
        - USER
        - ADMIN
...
  /api/trips/{tripId}/path:
    get:
      operationId: TripWorkspaceDataResource_getTripPath
      tags:
      - "User: Trips"
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: endTime
        in: query
        schema:
          type: string
      - name: startTime
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Get Trip Path
      security:
      - jwt:
        - USER
        - ADMIN
...
  /api/trips/{tripId}/timeline:
    get:
      operationId: TripWorkspaceDataResource_getTripTimeline
      tags:
      - "User: Trips"
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: endTime
        in: query
        schema:
          type: string
      - name: startTime
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Get Trip Timeline
      security:
      - jwt:
        - USER
        - ADMIN
...
info:
  title: GeoPulse API
  version: 1.34.1
  description: GeoPulse backend REST API documentation.
  contact:
    name: tess1o
    email: kerriden1@gmail.com
servers:
- url: http://localhost:8080